+++ CHANGES.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix wrong team exported as winning team
 * Fix island not being saved correctly when game ends
 * Save a few bytes when saving won islands
+* Fix overlapping text in level select screens when the resolution is low
 
 ### Removed